refactor(components): move ToggleableInput from Primitive to Generic
