+++ src/components/Generic/index.ts
@@ -4,6 +4,7 @@ export * from "./Section/Section"
 export * from "./SocialLinks/SocialLinks"
 export * from "./SponsorBadge/SponsorBadge"
 export * from "./SponsorTicker/SponsorTicker"
+export * from "./ToggleableInput/ToggleableInput"
 export * from "./ValueCard/ValueCard"
 export * from "./ValuesAccordion/ValuesAccordion"
 export * from "./ValuesFingerprint/ValuesFingerprint"
