+++ src/components/Primitive/index.ts
@@ -11,4 +11,3 @@ export * from "./PinInput/PinInput"
 export * from "./Radio/Radio"
 export * from "./Select/Select"
 export * from "./Toast/Toast"
-export * from "./ToggleableInput/ToggleableInput"
